requirements.txt

fastapi==0.115.11
uvicorn==0.34.0
SQLAlchemy
python-dotenv==1.0.0
langchain==0.3.20
langchain-openai==0.3.8
langchain-community==0.3.19
langchain-core==0.3.44
faiss-cpu==1.10.0
openai==1.64.0
